requirements.txt
absl-py==1.2.0
astor==0.8.1
atari-py==0.2.6
box2d-py==2.3.8
cachetools==4.2.4
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
cloudpickle==1.6.0
cycler==0.11.0
Cython==0.29.31
decorator==5.1.1
dm-tree==0.1.7
fasteners==0.17.3
fonttools==4.37.1
gast==0.2.2
gin-config==0.1.3
glfw==2.5.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gym==0.19.0
h5py==3.7.0
idna==3.3
imageio==2.20.0
importlib-metadata==4.12.0
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mujoco-py==2.1.2.14
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
six==1.16.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gan==2.0.0
tensorflow-gpu==2.0.1
tensorflow-hub==0.12.0
tensorflow-probability==0.8.0
termcolor==1.1.0
tf-agents-nightly==0.2.0.dev20191125
tfp-nightly==0.9.0.dev20191125
threadpoolctl==3.1.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.36.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wrapt==1.14.1
XlsxWriter==3.0.3
zipp==3.8.1
